Properly update the node on shortpoll interval
The poll() and query() methods weren't doing anything, or at least
not what they are supposed to do.

poll() should poll the device, get updated values and update the
node drivers with those values.

query() reads the current device values and updates the node drivers
with those values.

The main difference is that poll() is called automatically, at the
polling interval and query() is called when the ISY requests updated
values (either by the user clicking the query button or via a program)